refact: move TextField to 'text' package
